refactor: replace Store type with History for improved clarity and consistency
